// src/BindingChaos.Web/src/api/models/SuggestedActionViewModel.ts
/* tslint:disable */
/* eslint-disable */
/**
 * BindingChaos.Web.Gateway
 * BindingChaos Web Gateway API
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * View model for a suggested action on a signal.
 * @export
 * @interface SuggestedActionViewModel
 */
export interface SuggestedActionViewModel {
    /**
     * The unique identifier of the suggested action.
     * @type {string}
     * @memberof SuggestedActionViewModel
     */
    id?: string;
    /**
     * The action type code name (e.g. MakeACall, VisitAWebpage).
     * @type {string}
     * @memberof SuggestedActionViewModel
     */
    actionType?: string;
    /**
     * The phone number to call. Present for MakeACall actions.
     * @type {string}
     * @memberof SuggestedActionViewModel
     */
    phoneNumber?: string | null;
    /**
     * The URL to visit. Present for VisitAWebpage actions.
     * @type {string}
     * @memberof SuggestedActionViewModel
     */
    url?: string | null;
    /**
     * Optional free-text context provided by the suggester.
     * @type {string}
     * @memberof SuggestedActionViewModel
     */
    details?: string | null;
    /**
     * The pseudonym of the participant who suggested the action.
     * @type {string}
     * @memberof SuggestedActionViewModel
     */
    suggestedByPseudonym?: string;
    /**
     * When the action was suggested, in ISO 8601 format.
     * @type {string}
     * @memberof SuggestedActionViewModel
     */
    suggestedAt?: string;
}

/**
 * Check if a given object implements the SuggestedActionViewModel interface.
 */
export function instanceOfSuggestedActionViewModel(value: object): value is SuggestedActionViewModel {
    return true;
}

export function SuggestedActionViewModelFromJSON(json: any): SuggestedActionViewModel {
    return SuggestedActionViewModelFromJSONTyped(json, false);
}

export function SuggestedActionViewModelFromJSONTyped(json: any, ignoreDiscriminator: boolean): SuggestedActionViewModel {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'actionType': json['actionType'] == null ? undefined : json['actionType'],
        'phoneNumber': json['phoneNumber'] == null ? undefined : json['phoneNumber'],
        'url': json['url'] == null ? undefined : json['url'],
        'details': json['details'] == null ? undefined : json['details'],
        'suggestedByPseudonym': json['suggestedByPseudonym'] == null ? undefined : json['suggestedByPseudonym'],
        'suggestedAt': json['suggestedAt'] == null ? undefined : json['suggestedAt'],
    };
}

export function SuggestedActionViewModelToJSON(json: any): SuggestedActionViewModel {
    return SuggestedActionViewModelToJSONTyped(json, false);
}

export function SuggestedActionViewModelToJSONTyped(value?: SuggestedActionViewModel | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'actionType': value['actionType'],
        'phoneNumber': value['phoneNumber'],
        'url': value['url'],
        'details': value['details'],
        'suggestedByPseudonym': value['suggestedByPseudonym'],
        'suggestedAt': value['suggestedAt'],
    };
}